Se agrego las alertas

+++ Proyecto Integrado/ProyectoIntegradorFinal/src/app/components/main/CategoryProduct/list-category-product/list-category-product.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { CategoryProduct } from '../../../../Models/product';
 import { CategoryProductService } from '../../../../Service/category-product.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-list-category-product',
@@ -23,13 +24,25 @@ export class ListCategoryProductComponent {
     });
   }
 
-  deleteCategory(id: number) {
-    alert('Eliminando la categoria' + id);
-    this.categoryProductService
-      .deleteCategoriesProduct(id)
-      .subscribe((data) => {
-        console.log('CAMBIANDO EL DELETE', data);
-        this.getListCategory();
-      });
+  deleteCategory(id: number, category: string) {
+    Swal.fire({
+      title: 'Eliminar Categoría',
+      text: `¿Estás seguro de que deseas eliminar el categoría ${category}?`,
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Sí, eliminarlo',
+      cancelButtonText: 'Cancelar',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.categoryProductService
+          .deleteCategoriesProduct(id)
+          .subscribe((data) => {
+            console.log('CAMBIANDO EL DELETE', data);
+            this.getListCategory();
+          });
+      }
+    });
   }
 }
