+++ Proyecto Integrado/ProyectoIntegradorFinal/src/app/components/main/CategorySupplier/list-category-supplier/list-category-supplier.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { CategorySupplier } from '../../../../Models/supplier';
 import { CategorySupplierService } from '../../../../Service/category-supplier.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-list-category-supplier',
@@ -23,13 +24,25 @@ export class ListCategorySupplierComponent {
     });
   }
 
-  deleteCategory(id: number) {
-    alert('Eliminando el rubro' + id);
-    this.categorySupplierService
-      .deleteCategoriesSupplier(id)
-      .subscribe((data) => {
-        console.log('CAMBIANDO EL DELETE', data);
-        this.getListCategory();
-      });
+  deleteCategory(id: number, category: string) {
+    Swal.fire({
+      title: 'Eliminar Rubro',
+      text: `¿Estás seguro de que deseas eliminar el rubro ${category}?`,
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Sí, eliminarlo',
+      cancelButtonText: 'Cancelar',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.categorySupplierService
+          .deleteCategoriesSupplier(id)
+          .subscribe((data) => {
+            console.log('CAMBIANDO EL DELETE', data);
+            this.getListCategory();
+          });
+      }
+    });
   }
 }
